move token swap to ui component
